Enhancement: Extract AbstractTestCase
